apps/session-agent/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "foss-browserlab-session-agent"
version = "0.1.0"
description = "Chromium session worker for foss-browserlab"
requires-python = ">=3.12"
dependencies = [
  "aiortc>=1.9.0,<2.0.0",
  "av>=12.3.0,<13.0.0",
  "httpx>=0.28.1,<1.0.0",
  "mss>=10.0.0,<11.0.0",
  "numpy>=2.2.1,<3.0.0",
  "Pillow>=12.1.1,<13.0.0",
  "websockets>=15.0.1,<16.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0"
]

[project.scripts]
browserlab-session-agent = "session_agent.main:main"
browserlab-session-screenshot = "session_agent.screenshot:main"

[tool.hatch.build.targets.wheel]
packages = ["session_agent"]
